+++ ConnectFramework/ConnectFramework/Helpers/DataHelper.swift
@@ -1,0 +1,32 @@
+//
+//  DataHelper.swift
+//  ConnectFramework
+//
+//  Created by Konstantin Yurchenko, Jr on 9/12/24.
+//
+
+import Foundation
+
+public func load<T: Decodable>(_ filename: String) -> T {
+    let data: Data
+    
+    guard let file = Bundle(identifier: "ai.prorobot.ConnectFramework")?.url(forResource: filename, withExtension: nil)
+    else {
+        fatalError("Couldn't find \(filename) in main bundle.")
+    }
+
+
+    do {
+        data = try Data(contentsOf: file)
+    } catch {
+        fatalError("Couldn't load \(filename) from main bundle:\n\(error)")
+    }
+
+
+    do {
+        let decoder = JSONDecoder()
+        return try decoder.decode(T.self, from: data)
+    } catch {
+        fatalError("Couldn't parse \(filename) as \(T.self):\n\(error)")
+    }
+}
