+++ SkatePay/SkatePay/AppData.swift
@@ -6,34 +6,11 @@
 //
 
 import Foundation
+import ConnectFramework
 
 @Observable
 class AppData {
     var landmarks: [Landmark] = load("landmarkData.json")
     var users: [User] = load("userData.json")
     var profile = Profile.default
-}
-
-func load<T: Decodable>(_ filename: String) -> T {
-    let data: Data
-
-    guard let file = Bundle.main.url(forResource: filename, withExtension: nil)
-    else {
-        fatalError("Couldn't find \(filename) in main bundle.")
-    }
-
-
-    do {
-        data = try Data(contentsOf: file)
-    } catch {
-        fatalError("Couldn't load \(filename) from main bundle:\n\(error)")
-    }
-
-
-    do {
-        let decoder = JSONDecoder()
-        return try decoder.decode(T.self, from: data)
-    } catch {
-        fatalError("Couldn't parse \(filename) as \(T.self):\n\(error)")
-    }
 }
