--- SkatePay/SkatePay/Resources/userData.json
@@ -1,29 +1,0 @@
-[
-    {
-        "id": 1001,
-        "name": "HQ",
-        "npub": "npub1ydcksr7z0a2mk0fnhqkfd0dkgapdgqg2l39mfcuwwwuaeuf6r6qqzq7z7v",
-        "solanaAddress": "",
-        "relayUrl": "wss://relay.primal.net",
-        "imageName": "user-hub",
-        "isFavorite": true
-    },
-    {
-        "id": 1002,
-        "name": "Support",
-        "npub": "npub1uxp7mwl2mtetc4qmr0y6ck0p0y50c3zhglzzwvvdzf6dvpsjtvvq9gs05r",
-        "solanaAddress": "",
-        "relayUrl": "wss://relay.primal.net",
-        "imageName": "user-prorobot",
-        "isFavorite": false
-    },
-    {
-        "id": 1003,
-        "name": "dogtown",
-        "npub": "npub1hq5ae8m5vz3ys6rjg8c35z02xr80sfh9lyzqw69utrp5lryflvjqeqagjc",
-        "solanaAddress": "",
-        "relayUrl": "wss://relay.damus.io",
-        "imageName": "user-skatepay",
-        "isFavorite": false
-    }
-]
